Worked out a demo for panning the visible screen over the tile map.
Worked out functions to set the dimensions of the tile map width and height.

When in conio, scrolling is disabled (by calling scroll(0)), then printf will print beyond the visible screen borders and will print until the tile map border is reached.

I still need to add the scrolling functions, but that is not too much.

I have stolen your irq routine to pan the screen smoothly!

When all the characters are drawn, you can still type or move the cursor :-).